Create event package for client and consolidate events
