+++ src/main/java/org/protege/editor/owl/client/action/ShowUncommittedChangesAction.java
@@ -16,10 +16,10 @@ import javax.swing.JOptionPane;
 import javax.swing.KeyStroke;
 
 import org.protege.editor.core.ui.util.JOptionPaneEx;
-import org.protege.editor.owl.client.ClientSessionChangeEvent;
-import org.protege.editor.owl.client.ClientSessionListener;
-import org.protege.editor.owl.client.ClientSessionChangeEvent.EventCategory;
 import org.protege.editor.owl.client.api.exception.SynchronizationException;
+import org.protege.editor.owl.client.event.ClientSessionChangeEvent;
+import org.protege.editor.owl.client.event.ClientSessionListener;
+import org.protege.editor.owl.client.event.ClientSessionChangeEvent.EventCategory;
 import org.protege.editor.owl.client.ui.UncommittedChangesPanel;
 import org.protege.editor.owl.client.util.ClientUtils;
 import org.protege.editor.owl.model.OWLWorkspace;
