+++ src/main/java/org/protege/editor/owl/client/diff/model/LogDiffManager.java
@@ -6,7 +6,7 @@ import org.protege.editor.owl.OWLEditorKit;
 import org.protege.editor.owl.client.ClientSession;
 import org.protege.editor.owl.client.diff.DiffFactory;
 import org.protege.editor.owl.client.diff.DiffFactoryImpl;
-import org.protege.editor.owl.client.diff.ui.CommitOperationListener;
+import org.protege.editor.owl.client.event.CommitOperationListener;
 import org.protege.editor.owl.model.OWLModelManager;
 import org.protege.editor.owl.model.event.EventType;
 import org.protege.editor.owl.model.event.OWLModelManagerListener;
